iface: Prefer link local IPv6 address over all other ones
When choosing the IPv6 address to store for an `iface` and subsequently for
a `mdnsd` instance, the algorithm mimics the one for IPv4. This is wrong.
According to RFC 6724 the selection algorithm for the source address
selection shall prefer addresses in the appropriate scope for the destination
address. In our case the destination address is `ff02::fb` which means a
multicast address with *link local* scope. So the source address that the
OS will choose will be the closest scope to the link local scope.
The algorithm selecting the `iface` IPv6 address is changed accordingly
to prefer the smallest scope instead of the largest one.