libmdnsd: Pass source address of originating RR in conflict callback
When a resource record is received with conflicting data, the library
checks if it comes from its own interface (IP address), otherwise a
callback is called with the conflicting record. But if the host has
multiple interfaces, the packet may have been looped back from a different
interface, meaning this is not a conflict. The mdnsd instance only knows
one interface. To detect this kind of non-conflicts, the host needs the
information where the packet came from in the callback.

This patch adds a `sockaddr` parameter to the conflict callback, so that
the application can determine if it is a conflict or not.